Move to package cook.scheduler.client
